+++ CHANGELOG.md
@@ -2,12 +2,15 @@
 
 ## [Unreleased]
 ### Added
+- type-guards have been reworked from the ground up (#516)
 - `TypeGuard` is retained in inferred types (#504)
 - Type narrowing is applied from lambda execution (#504)
 - `--ide` flag (#501)
 ### Enhancements
 - `show-error-context`/`pretty` are now on by default (#501)
 - Show fake column number when `--show-error-end` (#501)
+- Error messages point to basedmypy docs (#516)
+- `Callable` types in error messages don't contain `mypy_extensions` (#516)
 ### Fixes
 - Don't show "X defined here" when error context is hidden (#498)
 - Fix issue with reveal code in ignore message (#490)